# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: P2P Exchange Ethiopia API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: tab
          in: query
          required: false
          schema:
            type: string
            enum: [ongoing, fulfilled]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [unpaid, paid, completed, cancelled, appeal]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order from an ad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/mark-paid:
    post:
      operationId: markOrderPaid
      tags: [orders]
      summary: Buyer marks order as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/release:
    post:
      operationId: releaseCrypto
      tags: [orders]
      summary: Seller releases crypto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelOrderInput"
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/appeal:
    post:
      operationId: raiseAppeal
      tags: [orders]
      summary: Raise a dispute appeal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppealInput"
      responses:
        "201":
          description: Appeal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appeal"
  /orders/{id}/feedback:
    post:
      operationId: submitFeedback
      tags: [orders]
      summary: Submit feedback for completed order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"

  # Messages
...
  /messages/{orderId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{orderId}/read:
    post:
      operationId: markMessagesRead
      tags: [messages]
      summary: Mark messages as read
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read

  # Profile
...
  /admin/kyc:
    get:
      operationId: adminListKyc
      tags: [admin]
      summary: List all KYC submissions (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, verified, rejected, more_info_required]
      responses:
        "200":
          description: KYC submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KycSubmission"
  /admin/kyc/{userId}:
    get:
      operationId: adminGetKyc
      tags: [admin]
      summary: Get KYC submission for a user (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: KYC submission detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycSubmission"
  /admin/kyc/{userId}/review:
    post:
      operationId: adminReviewKyc
      tags: [admin]
      summary: Review a KYC submission (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycReviewInput"
      responses:
        "200":
          description: Updated KYC status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycSubmission"
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        adId:
          type: integer
        buyerId:
          type: integer
        sellerId:
          type: integer
        buyerUsername:
          type: string
        sellerUsername:
          type: string
        amountUsdt:
          type: string
        amountEtb:
          type: string
        price:
          type: string
        paymentMethod:
          type: string
        paymentTimeLimit:
          type: integer
        status:
          type: string
          enum: [unpaid, paid, completed, cancelled, appeal]
        cancelReason:
          type: ["string", "null"]
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        sellerAccountName:
          type: ["string", "null"]
        sellerAccountNumber:
          type: ["string", "null"]
      required:
        - id
        - adId
        - buyerId
        - sellerId
        - buyerUsername
        - sellerUsername
        - amountUsdt
        - amountEtb
        - price
        - paymentMethod
        - paymentTimeLimit
        - status
        - unreadCount
        - createdAt
    OrderInput:
      type: object
      properties:
        adId:
          type: integer
        amountUsdt:
          type: string
        amountEtb:
          type: string
        paymentMethod:
          type: string
      required:
        - adId
        - amountUsdt
        - amountEtb
        - paymentMethod
    CancelOrderInput:
      type: object
      properties:
        reason:
          type: string
    AppealInput:
      type: object
      properties:
        reason:
          type: string
        description:
          type: string
        evidenceUrls:
          type: array
          items:
            type: string
      required:
        - reason
        - description
    Appeal:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        raisedBy:
          type: integer
        reason:
          type: string
        description:
          type: string
        evidenceUrls:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, resolved]
        adminDecision:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - raisedBy
        - reason
        - description
        - evidenceUrls
        - status
        - createdAt
    FeedbackInput:
      type: object
      properties:
        type:
          type: string
          enum: [positive, negative]
        comment:
          type: string
      required:
        - type
    Feedback:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        type:
          type: string
          enum: [positive, negative]
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - fromUserId
        - toUserId
        - type
        - createdAt
    Message:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        senderId:
          type: integer
        senderUsername:
          type: string
        receiverId:
          type: integer
        content:
          type: string
        type:
          type: string
          enum: [text, image, system]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - senderId
        - senderUsername
        - receiverId
        - content
        - type
        - isRead
        - createdAt
    MessageInput:
      type: object
      properties:
        content:
          type: string
        type:
          type: string
          enum: [text, image]
      required:
        - content
        - type
...
    KycSubmission:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        dateOfBirth:
          type: string
        nationality:
          type: string
        idType:
          type: string
        frontImageUrl:
          type: string
        backImageUrl:
          type: ["string", "null"]
        selfieUrl:
          type: string
        livenessResult:
          type: object
        status:
          type: string
          enum: [pending, verified, rejected, more_info_required]
        rejectionReason:
          type: ["string", "null"]
        adminMessage:
          type: ["string", "null"]
        reviewedBy:
          type: ["integer", "null"]
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: ["string", "null"]
      required:
        - id
        - userId
        - username
        - email
        - fullName
        - dateOfBirth
        - nationality
        - idType
        - frontImageUrl
        - selfieUrl
        - livenessResult
        - status
        - submittedAt
    KycReviewInput:
      type: object
      properties:
        decision:
          type: string
          enum: [verified, rejected, more_info_required]
        rejectionReason:
          type: string
        adminMessage:
          type: string
      required:
        - decision
...