# lib/api-zod/src/generated/types/appealInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * P2P Exchange Ethiopia API
 * OpenAPI spec version: 0.1.0
 */

export interface AppealInput {
  reason: string;
  description: string;
  evidenceUrls?: string[];
}