# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * P2P Exchange Ethiopia API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackType } from './feedbackType';

export interface Feedback {
  id: number;
  orderId: number;
  fromUserId: number;
  toUserId: number;
  type: FeedbackType;
  /** @nullable */
  comment?: string | null;
  createdAt: Date;
}