# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * P2P Exchange Ethiopia API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  adId: number;
  buyerId: number;
  sellerId: number;
  buyerUsername: string;
  sellerUsername: string;
  amountUsdt: string;
  amountEtb: string;
  price: string;
  paymentMethod: string;
  paymentTimeLimit: number;
  status: OrderStatus;
  /** @nullable */
  cancelReason?: string | null;
  unreadCount: number;
  createdAt: Date;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  sellerAccountName?: string | null;
  /** @nullable */
  sellerAccountNumber?: string | null;
}